Fix integer overflow in xmhf_baseplatform_arch_x86_udelay()

+++ xmhf/src/xmhf-core/xmhf-runtime/xmhf-baseplatform/arch/x86/bplt-x86-pit.c
@@ -69,7 +69,7 @@ void xmhf_baseplatform_arch_x86_udelay(u32 usecs){
   outb(0xB0, 0x43);
   
   //compute appropriate latch register value depending on usecs
-  latchregval = (1193182 * usecs) / 1000000;
+  latchregval = ((u64)1193182 * usecs) / 1000000;
 
   //write latch register to ch-2
   val = (u8)latchregval;
